tp-runner: Bash script to run the tests.
- Deal with Runtime Error and Time Limit Exceeded
While trying to implement these feature I realized that the system()
function is not enough, we need more control over the execution of the
program and I think that a bash script is OK, at least for now.

tp-runner is a minimal bash script whose purpose is to run the program
and return appropriate codes.

- Change .test_info structure
The current structure of the .test_info file is: In a single line two
parameters, the number of test cases and the name of the executable
file.

From now the information related to each test case uses the form
key = value, which is more comprehensible and allows us to manage more
settings easily and without a particular order in the file.

These are the available settings for now:
    tests       => The number of test cases.
    src_file    => Source file.
    lang        => Programming language in which the solution is implemented.
    time_limit  => Maximum amount of time our program can run.